move the ExportJob into the import dir
